+++ js/data.js
@@ -1,0 +1,54 @@
+/* ----------------- Генерация данных ----------------- */
+
+const TITLES = [
+  'Выгодное предложение',
+  'Не упустите свой шанс',
+  'Горячая скидка',
+  'Цены снижены',
+  'Пора бронировать',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTION = [
+  'Квартирка что надо',
+  'Лучших апартаментов просто не бывает',
+  'Здесь есть даже горячая вода',
+  'Тихое уютное место и без соседей',
+  'Да, маленькая. Зато уютная.',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {TITLES, TYPES, CHECKIN, CHECKOUT, FEATURES, DESCRIPTION, PHOTOS};
