+++ js/main.js
@@ -1,98 +1,7 @@
-// Функция, возвращающая случайное целое число из переданного диапазона включительно
-
-const getRandomIntFromRange = function (from, to) {
-  if (from < 0 || to <= from) {
-    return;
-  }
-  from = Math.ceil(from);
-  to = Math.floor(to);
-  return Math.floor(Math.random() * (to - from + 1)) + from;
-};
-
-getRandomIntFromRange(5, 15);
-
-// Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно
-
-const getRandomFloatRange = function (from, to, count) {
-  if (from < 0 || to <= from) {
-    return;
-  }
-
-  return Number((Math.random() * (to - from) + from).toFixed(count));
-};
-
-getRandomFloatRange(1.2, 3.4, 2);
-
-/* ----------------- Генерация данных ----------------- */
-
-const TITLES = [
-  'Выгодное предложение',
-  'Не упустите свой шанс',
-  'Горячая скидка',
-  'Цены снижены',
-  'Пора бронировать',
-];
-
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-
-const CHECKIN = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const CHECKOUT = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const DESCRIPTION = [
-  'Квартирка что надо',
-  'Лучших апартаментов просто не бывает',
-  'Здесь есть даже горячая вода',
-  'Тихое уютное место и без соседей',
-  'Да, маленькая. Зато уютная.',
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-// Данную функцию getArray честно подсмотрел в интернете
-const getArray = (data) => {
-  const maxLength = data.length;
-  const lengthOfArray = getRandomIntFromRange(1, maxLength);
-  const array = [];
-
-  while ( array.length < lengthOfArray) {
-    const indexOfEl = getRandomIntFromRange(0, maxLength - 1);
-    const el = data[indexOfEl];
-
-    if (!array.includes(el)) {
-      array.push(el);
-    }
-  }
-
-  return array;
-};
+import {getRandomIntFromRange} from './random-int.js';
+import {getRandomFloatRange} from './random-float.js';
+import {getArray} from './get-array.js';
+import {TITLES, TYPES, CHECKIN, CHECKOUT, FEATURES, DESCRIPTION, PHOTOS} from './data.js';
 
 const createDataGeneration = () => {
   const randomNumberPhoto = getRandomIntFromRange(1, 10);
@@ -142,4 +51,5 @@ const createDataGeneration = () => {
 };
 
 const similarObjects = Array.from({length: 10}, createDataGeneration);
+
